+++ CHANGELOG.md
@@ -4,6 +4,9 @@
   bundled apworld. See the documentation of `goal` in `Dark Souls III Options
   Template.yaml` for details.
 
+* Add an option to customize death link amnesty. See the documentation of
+  `death_link_amnesty` in `Dark Souls III Options Template.yaml` for details.
+
 ## 4.0.0-alpha.8
 
 * Fix a bug where the client would send empty `CreateHints` packets over and